Rename UncoupledStatement to index for neat imports
